+++ src/mpd_client/playlists.h
@@ -19,7 +19,7 @@ enum playlist_types {
 time_t mpd_client_get_playlist_mtime(struct t_partition_state *partition_state, const char *playlist);
 bool mpd_client_playlist_clear(struct t_partition_state *partition_state, const char *plist, sds *error);
 bool mpd_client_playlist_shuffle(struct t_partition_state *partition_state, const char *uri, sds *error);
-bool mpd_client_playlist_sort(struct t_partition_state *partition_state, const char *uri, const char *tagstr, sds *error);
+bool mpd_client_playlist_sort(struct t_partition_state *partition_state, const char *uri, const char *tagstr, bool sortdesc, sds *error);
 long mpd_client_enum_playlist(struct t_partition_state *partition_state, const char *playlist, bool empty_check);
 long mpd_client_playlist_validate(struct t_partition_state *partition_state, const char *playlist, bool remove, sds *error);
 long mpd_client_playlist_validate_all(struct t_partition_state *partition_state, bool remove, sds *error);